requirements.txt
flask==3.0.0
flask-cors==4.0.0
tqdm==4.66.1
wget==3.2
pandas==2.1.4
numpy==1.26.3
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.0
redis==5.0.1
requests==2.31.0
certifi==2023.7.22
gunicorn==21.2.0
